fix: remove old slider and update imports
