+++ components/paragraphs/MaterialSlider/MaterialSlider.tsx
@@ -1,9 +1,14 @@
+"use client"
+
+import { useWindowSize } from "@uidotdev/usehooks"
+import "keen-slider/keen-slider.min.css"
+import { KeenSliderOptions, KeenSliderPlugin, useKeenSlider } from "keen-slider/react"
 import Link from "next/link"
-import React, { useEffect, useRef, useState } from "react"
+import React, { useEffect, useState } from "react"
 
 import { Button } from "@/components/shared/button/Button"
 import Icon from "@/components/shared/icon/Icon"
-import { WorkCardEmpty, WorkCardSkeleton } from "@/components/shared/workCard/WorkCard"
+import WorkCard, { WorkCardEmpty, WorkCardSkeleton } from "@/components/shared/workCard/WorkCard"
 import WorkCardWithCaption from "@/components/shared/workCard/WorkCardWithCaption"
 import {
   ParagraphGoMaterialSliderAutomatic,
@@ -18,103 +23,182 @@ type MaterialSliderProps = {
   title: ParagraphGoMaterialSliderAutomatic["title"] | ParagraphGoMaterialSliderManual["title"]
 }
 
-const MaterialSlider = ({ works, title }: MaterialSliderProps) => {
-  const sliderRef = useRef<HTMLDivElement | null>(null)
-  const [areButtonsDisabled, setAreButtonsDisabled] = useState(false)
-  const [canScrollLeft, setCanScrollLeft] = useState(false)
-  const [canScrollRight, setCanScrollRight] = useState(true)
+const defaultSliderOptions: KeenSliderOptions = {
+  initial: 0,
+  slides: {
+    origin: "auto",
+    spacing: 12,
+    perView: 1.1,
+  },
+  breakpoints: {
+    "(min-width: 768px)": {
+      slides: {
+        origin: "auto",
+        spacing: 12,
+        perView: () => {
+          return 2.1
+        },
+      },
+    },
+    "(min-width: 1024px)": {
+      slides: {
+        spacing: 24,
+        perView: () => {
+          return 3
+        },
+      },
+    },
+  },
+}
 
-  const updateScrollButtons = () => {
-    if (!sliderRef.current) return
-    const { scrollLeft, scrollWidth, clientWidth } = sliderRef.current
-    setCanScrollLeft(scrollLeft > 0)
-    setCanScrollRight(scrollLeft + clientWidth < scrollWidth)
+const WheelControls: KeenSliderPlugin = slider => {
+  let touchTimeout: ReturnType<typeof setTimeout>
+  let position: {
+    x: number
+    y: number
+  }
+  let wheelActive: boolean
+
+  function dispatch(e: WheelEvent, name: string) {
+    position.x -= e.deltaX
+    position.y -= e.deltaY
+    slider.container.dispatchEvent(
+      new CustomEvent(name, {
+        detail: {
+          x: position.x,
+          y: position.y,
+        },
+      })
+    )
   }
 
-  useEffect(() => {
-    updateScrollButtons()
-  }, [works])
-
-  const scroll = (direction: "left" | "right") => {
-    if (!sliderRef.current) return
-    const scrollAmount = sliderRef.current.clientWidth * 0.9
-    const finalScrollAmount = direction === "left" ? -scrollAmount : scrollAmount
-    setAreButtonsDisabled(true)
-    sliderRef.current.scrollBy({ left: finalScrollAmount, behavior: "smooth" })
-    setTimeout(() => {
-      updateScrollButtons()
-      setAreButtonsDisabled(false)
-    }, 500)
+  function wheelStart(e: WheelEvent) {
+    position = {
+      x: e.pageX,
+      y: e.pageY,
+    }
+    dispatch(e, "ksDragStart")
   }
 
+  function wheel(e: WheelEvent) {
+    dispatch(e, "ksDrag")
+  }
+
+  function wheelEnd(e: WheelEvent) {
+    dispatch(e, "ksDragEnd")
+  }
+
+  function eventWheel(e: WheelEvent) {
+    if (e.deltaX === 0) {
+      return
+    }
+
+    e.preventDefault()
+    if (!wheelActive) {
+      wheelStart(e)
+      wheelActive = true
+    }
+    wheel(e)
+    clearTimeout(touchTimeout)
+    touchTimeout = setTimeout(() => {
+      wheelActive = false
+      wheelEnd(e)
+    }, 50)
+  }
+
+  slider.on("created", () => {
+    slider.container.addEventListener("wheel", eventWheel, {
+      passive: false,
+    })
+  })
+}
+
+const MaterialSliderNew = ({ works, title }: MaterialSliderProps) => {
+  const [sliderRef, internalSlider] = useKeenSlider(defaultSliderOptions, [WheelControls])
+  const [reachedStart, setReachStart] = useState(true)
+  const [reachedEnd, setReachEnd] = useState(true)
+  const size = useWindowSize()
+
   useEffect(() => {
-    const slider = sliderRef.current
-    if (!slider) return
-    const handleScroll = () => {
-      updateScrollButtons()
-    }
-    slider.addEventListener("scroll", handleScroll)
-    return () => {
-      slider.removeEventListener("scroll", handleScroll)
-    }
+    updateSlidePosition()
+
+    internalSlider.current?.on("slideChanged", () => {
+      updateSlidePosition()
+    })
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
+  // If window size or children amount changes, update the slider
+  useEffect(() => {
+    internalSlider.current?.update()
+  }, [size.width, internalSlider])
+
+  const updateSlidePosition = () => {
+    setReachStart(internalSlider.current?.track.details.rel === 0)
+    setReachEnd(
+      internalSlider.current?.track.details.maxIdx === internalSlider.current?.track.details.rel
+    )
+  }
+
+  const onLeftClick = () => {
+    internalSlider.current?.prev()
+  }
+
+  const onRightClick = () => {
+    internalSlider.current?.next()
+  }
+
   return (
-    <div className="bg-background-overlay">
+    <div className="bg-background-overlay overflow-hidden">
       <div className="content-container grid-go">
-        <h2 className="text-typo-heading-2 col-span-full my-12">{title}</h2>
-        <hr className="col-span-full mb-6" />
-        {(canScrollLeft || canScrollRight) && (
-          <div className="gap-grid-gap col-span-full flex flex-row justify-end">
+        <div className="pt-paragraph-spacing col-span-full flex items-center justify-between">
+          <h2 className="text-typo-heading-2">{title}</h2>
+          <div className="flex flex-row justify-end gap-x-4">
             <Button
-              onClick={() => {
-                scroll("left")
-              }}
+              disabled={reachedStart}
               variant="icon"
               ariaLabel="Vis forrige værker"
-              disabled={!canScrollLeft || areButtonsDisabled}>
+              onClick={() => onLeftClick()}>
               <Icon className="h-[24px] w-[24px]" name="arrow-left" />
             </Button>
             <Button
-              onClick={() => {
-                scroll("right")
-              }}
+              disabled={reachedEnd}
               variant="icon"
               ariaLabel="Vis næste værker"
-              disabled={!canScrollRight || areButtonsDisabled}>
+              onClick={() => onRightClick()}>
               <Icon className="h-[24px] w-[24px]" name="arrow-right" />
             </Button>
           </div>
-        )}
-        <div
-          ref={sliderRef}
-          className={"gap-grid-gap col-span-full flex flex-row overflow-x-scroll py-6 md:py-12"}>
-          {works ? (
-            works.map(work => (
-              <Link
-                key={work.workId}
-                aria-label={`Tilgå værket ${work.titles.full[0]} af ${displayCreators(work.creators, 1)}`}
-                className="focus-visible w-[70%] shrink-0 sm:w-[40%] lg:w-[30%]"
-                href={resolveUrl({
-                  routeParams: { work: "work", wid: work.workId },
-                  queryParams: {
-                    type: work.manifestations.bestRepresentation.materialTypes[0]
-                      .materialTypeGeneral.code,
-                  },
-                })}>
-                <WorkCardWithCaption
-                  work={work}
-                  classNameWorkCard={"bg-background dark-mode-transition w-[100%]"}
-                  className="max-w-[100%]"
-                  isWithTilt
-                />
-              </Link>
-            ))
-          ) : (
-            <div className="w-[70%] shrink-0 sm:w-[40%] lg:w-[30%]">
-              <WorkCardEmpty />
+        </div>
+
+        <div className="-mx-grid-edge px-grid-edge col-span-full">
+          <div className="my-paragraph-spacing">
+            <div ref={sliderRef} className={"keen-slider !overflow-visible"}>
+              {works ? (
+                works.map(work => (
+                  <Link
+                    key={work.workId}
+                    aria-label={`Tilgå værket ${work.titles.full[0]} af ${displayCreators(work.creators, 1)}`}
+                    className="keen-slider__slide focus-visible outline-accent-foreground focus:outline-offset-2"
+                    href={resolveUrl({
+                      routeParams: { work: "work", wid: work.workId },
+                      queryParams: {
+                        type: work.manifestations.bestRepresentation.materialTypes[0]
+                          .materialTypeGeneral.code,
+                      },
+                    })}>
+                    <WorkCardWithCaption title={work.titles.full[0]} creators={work.creators || []}>
+                      <WorkCard className="bg-background" work={work} isWithTilt={true} />
+                    </WorkCardWithCaption>
+                  </Link>
+                ))
+              ) : (
+                <div className="">
+                  <WorkCardEmpty />
+                </div>
+              )}
             </div>
-          )}
+          </div>
         </div>
       </div>
     </div>
@@ -144,4 +228,4 @@ export const MaterialSliderSkeleton = () => {
   )
 }
 
-export default MaterialSlider
+export default MaterialSliderNew
