+++ components/paragraphs/MaterialSlider/MaterialSliderAutomatic.tsx
@@ -5,7 +5,7 @@ import React from "react"
 import { ParagraphGoMaterialSliderAutomatic } from "@/lib/graphql/generated/dpl-cms/graphql"
 import { useComplexSearchForWorkTeaserQuery } from "@/lib/graphql/generated/fbi/graphql"
 
-import MaterialSliderNew, { MaterialSliderSkeleton } from "./MaterialSliderNew"
+import MaterialSlider, { MaterialSliderSkeleton } from "./MaterialSlider"
 
 type MaterialSliderAutomatic = {
   titleOptional: ParagraphGoMaterialSliderAutomatic["title"]
@@ -30,7 +30,7 @@ const MaterialSliderAutomatic = ({
 
   if (isLoading) return <MaterialSliderSkeleton />
 
-  return <MaterialSliderNew works={data?.complexSearch.works} title={titleOptional} />
+  return <MaterialSlider works={data?.complexSearch.works} title={titleOptional} />
 }
 
 export default MaterialSliderAutomatic
